A nfa transition now stores a char range instead of only a single char
This has the potential to make everything regarding constructing and
maybe minimizing a dfa more complicated, but allows us to easily add
primitive character classes like '.' or '[a-zA-Z0-9]'.

If we want to generate machine code later that will also work on ranges,
so why not work with ranges from the start?